requirements.txt
selenium
behave
pytest
allure-pytest
configparser
allure-behave
